requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
boto==2.49.0
branca==0.4.1
Brotli==1.0.9
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dash==1.16.2
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.1
django-bootstrap4==2.2.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-extensions==3.0.8
django-heroku==0.3.1
django-plotly-dash==1.4.2
djangorestframework==3.11.1
dpd-components==0.1.0
flake8==3.8.3
Flask==1.1.2
Flask-Compress==1.5.0
folium==0.11.0
future==0.18.2
gunicorn==19.0.0
idna==2.10
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
networkx==2.5
numpy==1.19.1
Pillow==7.2.0
plotly==4.10.0
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
pytz==2020.1
requests==2.24.0
retrying==1.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
gunicorn==19.0.0
